refactor app into a finely-controlled state machine
Types and functions are designed such that that app can only do what it's supposed to and only when it's supposed to do it.

For example, during a profile fetch the button is disabled not only to avoid excessive fetches, but because the view cannot even produce the Cmd to perform the fetch at that moment in time; The capability FetchCap is required to perform a fetch, and that's not available during the fetch.